[CHORE] Allow manual start of hash regeneration workflow (#3900)
The workflow now runs on a schedule, but that means we need to wait till midnight once a week to test it.

Add an `on: { workflow_dispatch: {} }` trigger to this workflow, so that it can be run whenever we want.

**How to test**

Merge this, go to the **Actions** tab, click on the workflow, then click **Run workflow**.