3839 [Testing] creating front end tests for features on the add students create new accounts page (#3845)
**Description**
This pull request adds test for all the features on the `for-teacher/create-accounts`  page. This includes tests for the following features:

- go back to class button
- remove row
- add rows
- generate passwords button
- reset button
- download login credentials
- create accounts button


**Fixes _issue or discussion number_**
Fixes #3839

**How to test**
These tests uses the accounts in the test database, to use this the database has to be fed first by running bash `feed_dev_database.sh` in the `hedy` (root) directory.
To run the tests on the command, do the the following: `npx cypress run --spec tests/cypress/e2e/for-teacher_page/class_page/create_accounts_page`
This tests all the create_accounts page tests.