[TESTING, CYPRESS] 3703 tests for saving code on /hedy page (#3780)
**Description**

This pull request adds tests for the save button and save & share button on the `/hedy` page when trying to save a level.

**Fixes _issue or discussion number_**
Fixes #3703

**How to test**

These tests uses the accounts in the test database, to use this the database has to be fed first by running bash `feed_dev_database.sh` in the hedy (root) directory.
To run the tests on the command, do the the following: `npx cypress run --spec tests/cypress/e2e/hedy_page`
This tests all the hedy page tests.