[Language] Adding error to catch index out of bounds errors (#3788)
**Description**

Adds an exception to the code generated after transpiling to catch index out of bounds errors.
Things to do:
- [x] Figure out a way to make the message translatable (`gettext` works inside hedy.py!)
- [x] Try to show which line and variable caused the error. (Can show the variable, not the line!)

**Fixes #3463**

**How to test**

Since this is added **after** transpiling, should I add an e2e test?

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Contains one of the PR categories in the name
- [ ] Describes changes in the format above
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!