[TESTING] 3562 part of classpage, add student by link and username  (#3682)
[TESTING, CYPRESS]


**Description**
Adds cypress front end tests for features on the `/for-teacher/class/[class id]` page. One test checks if the link to add a student to the class is copied to the clipboard. And the other tests checks if a student is added when you add a student by username. I tried to implemented in a dynamic fashion.

**Fixes _issue or discussion number_**
Fixes #3562

**How to test**

Run
`export ADMIN_USER=admin_user; python app.py` in your virtual environment. This ensures that there is an active admin_user and Hedy is running.

Then you can go to the /tests folder and run:

`npx cypress run --spec cypress/e2e/for-teacher_page/class_page`