[CHORE] Design: add a colophon block at the bottom (#3901)
Most websites have a block at the bottom for the "fine print", policy, getting in touch, and other meta-details about the website.

Add one to the front page of the new design, and remove the **Learn more** button from the top bar (where it is eating valuable space).

![image](https://user-images.githubusercontent.com/524162/210244640-e9094baf-9494-4b35-9181-e3fb31cf29ac.png)

**How to test**

Observe the new design. The contents can be changed, they are currently not translated.