[FIX] Design: next level block overlay hides button (#3940)
If the "next level" button is disabled because a teacher chooses to do so, the popup informing children was hiding the "Reply" button.

Move it down over the button itself so the blocks don't interfere.

Also make the Login screen a bit wider, the word "Gebruikersnaam" is too long for the current box.

## Before

![image](https://user-images.githubusercontent.com/524162/212660455-2fad2fc6-81b4-4b8a-8570-809f0b2ce091.png)

![image](https://user-images.githubusercontent.com/524162/212660581-2fce2f9d-6d81-41cc-9a66-8f0c0b8889f3.png)


## After

![image](https://user-images.githubusercontent.com/524162/212660511-162a5745-6fa3-477d-84aa-32db392de468.png)

![image](https://user-images.githubusercontent.com/524162/212660295-cb098a9b-d2e1-4c5d-834f-905528b61aaf.png)