[UI] Slide download and view (#4023)
**Description**

Adds a new table to the for teachers page, where teachers can see slides and also download them.

![image](https://user-images.githubusercontent.com/45865185/217730375-2a620226-9043-44e3-b634-a8c1af8911d1.png)

The zip with the slides has the following structure according to the Slides.com format:

* A lib folder, that's empty, since Slides.com ignores our css and javascript anyways.
* A folder with the same name as the zip, empty because there are no images in our slides.
* A index.html files where `iframes` are changed by anchor tags, since they are not imported correctly to slides.com.

An example of the import can be found here https://slides.com/jesuspelay/deck-e2545a/fullscreen

**Fixes #3924**


**How to test**

* If this is a UI change: _describe how to run your code to see it in action. See this https://github.com/Felienne/hedy/pull/880#issue-1016304308 for an example_
* If this is a language change: _add a few tests showing the difference_

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Contains one of the PR categories in the name
- [ ] Describes changes in the format above
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!