fix: dropdown-selected keyword language is not passed to JavaScript
The keyword language selected in the keyword switching dropdown
caused the `?keyword_language=` parameter to be added to the URL,
but this parameter did not ultimately make its way into
`g.keyword_lang`, which is the variable that gets passed to the
JavaScript.

So according to the JS, the keyword language was always English.

Fix that.