[FIX] Do not drop messages from 'generate-client-messages'
The way we do client side translations now, by using
`generate-client-messages.py` to translate them, `gettext` loses the
information that we were interested in those translations because there
isn't any source code anymore that contains `gettext("whatever_key")`.
It then proceeds to drop the message from the `.pot` and `.po` files.

To prevent that, make sure we call the `gettext` function in
`generate-client-messages.py` as well, and add it to the list of
traced source files, so that `gettext` can keep on finding those keys.

For some reason, doing this completely changed the order of keys
in the `messages.pot` file. I'm not sure why, but I'm leaving it as is.