[UI IMPROVEMENT] Smoothen dropdown menus (#4069)
**Description**
Slide toggle hide a dropdown when the button is pressed again instead of the hacky re-slide open we currently have.

**Fixes**
This PR fixes #4066

**How to test**
Verify that the language dropdown, user profile dropdown and both the cheatsheet and language switcher dropdown all close when the button is pressed again. This makes the animation a lot more smooth than what currently happens when we press the button twice.