[FIX] Speech dropdown is gone (#4092)
The speech dropdown was being initialized outside of the `initialize` functions, and so depended on state that wasn't set yet (`theLanguage` was not initialized yet, and then the function did nothing).

Move the code into an explicit initializer and call it at the right time.

Add a test for it as well.