[FIX] Student needs to wait 5m or logout after joining class
The classes that a student is part of is cached on their session
(so we don't have to query the database on every page load).

However, we failed to refresh the cache when the user joins a class.
The result was that they had to wait 5 minutes for their session data to be
refreshed as usual, or log out and back in to force a refresh.

Instead, do the refresh on joining and leaving a class.

Thanks for catching this @TiBiBa!