[UI] Hide submit button on code page when not in class (#4156)
**Description**
The title should be self-explanatory, when the current user is not in a class, hide the submit button. Similarly we might want to hide the bar all together when a user is not logged in. But I'm not sure if that is desirable. I did find two new bugs when working on this, still have to make issues on them:

- When logged in and navigating to a join link, you are logged out (this should not be the case!)
- When joining a class the value on `current_user()` is not updated, the user is required to log out / log in again to set this right.

**Fixes**
This PR fixes #4154.