[FIX] Handle incomplete database rows (#4158)
Not all database rows have all columns they should have. The current code that was dealing with database rows was assuming they did.

Remove that assumption, change `r['name']` into `r.get('name', '')` so that the code doesn't throw.

Fixes #4153.

**How to test**

It's not clear how to reproduce this, it depends heavily on what's in the database.