[FIX] User doesn't need to wait anymore to submit programs (#4159)
The classes that a student is part of is cached on their session (so we don't have to query the database on every page load).

However, we failed to refresh the cache when the user joins a class. The result was that they had to wait 5 minutes for their session data to be refreshed as usual, or log out and back in to force a refresh.

Instead, do the refresh on joining and leaving a class.

Thanks for catching this @TiBiBa!

**How to test**

Invite a new student that's not in any classes. Join as that student, observe that you can submit programs immediately (instead of after 5 minutes)

(Fixing a bug mentioned in #4156)