[FIX] "Mark as teacher" error on admin page (#4174)
When marking a user as teacher, an error would be shown.

This is because I rewrote this code to use the (new) `postJson` function, but that function also expects a JSON response, and the `/admin/markAsTeacher` endpoint doesn't respond with anything.

Use `postNoResponse` instead.

**How to test**

On the admin page, mark a user as teacher. Observe that no error occurs.