[FIX] Student page can deal with builtin adventure being missing (#4176)
The old code was assuming that all built-in adventures that appear in the `ordered_adventures` actually still exist.

However, this list can also change and items can be removed from it, so deal with that by ignoring them.

**How to test**

Search for a class in the database that uses the `pressit` adventure, and log in as a student in that class.