[FIX] Do not overwrite programs with the same name
In the new autosave system, switching away from a tab automatically
saves your current program.

Some students are experiencing that (for some reason we don't understand
yet), their previous programs aren't automatically loaded, and then the
autosave overwrites their old program with a blank one.

This can only happen because we have a legacy bit of code in there,
which says that if you don't include a program id, it's going to
overwrite an old program with the same name. Since no one ever
changes the default save name, this is most likely the culprit of the
old program being overwritten (new empty program has the same name
as the old saved program, and is therefore clobbered).

Disable this behavior, as we don't need it any more and it in fact
behaves poorly in combination with autosave. Either we loaded
an old program and we overwrite it, or we didn't in which case
we should always save in a new "slot".