[FIX] Can load old saved programs on teacher adventures
Programs are saved with the adventure they were associated with. For
built-in adventures, this is the symbolic short name.  For teacher
adventures, we used to store them with the long display name, but
somewhere along the way this was changed to the adventure id.

The current code only uses the adventure id to select programs into
a tab, but the result of that is that programs that were saved before
the new standard cannot be loaded anymore.

Upon loading the student page, also identify programs by adventure name
if no programs can be found by adventure id.