[FIX] Restore loading old saved programs on teacher adventures (#4180)
Programs are saved with the adventure they were associated with. For built-in adventures, this is the symbolic short name.  For teacher adventures, we used to store them with the long display name, but somewhere along the way this was changed to the adventure id.

The current code only uses the adventure id to select programs into a tab, but the result of that is that programs that were saved before the new standard cannot be loaded anymore.

Upon loading the student page, also identify programs by adventure name if no programs can be found by adventure id.

**How to test**

Find a student in a class with custom adventures that were saved before the attribute name was changed. Log in as that student, observe that the program is loaded onto the customer adventure tab.