More work on the misconception card
Fixed the issue where it threw an error when clicking on a struggle item, had to do with passing jinja vars into a macro. Moved the block popup_content into the right location, otherwise it isn't drawn. 
Created a pop up card with transparent background elements to blend out the rest of the dashboard