fixed overlay height, fixed linking error
Added a height of 120% otherwise the transparent overlay did not cover all the content which made it look off. Fixed an error that crept in previously where I tried linking a student to the page already, but the functionality for that we will introduce at a later stage. 
Therefore it was error proofed in statistics.py