[FIX] Log session and user info for all HTTP requests (#4231)
We were not logging user-attributable information for all HTTP requests, only for a couple. This makes it very hard to debug things going wrong.

Make sure we log user-related information with every request (session and username), and emit a couple more bits of information whenever we end up saving a program so it'll be easier to debug weird saving issues.

**How to test**

Nothing to test.