[FIX] Scroll output to bottom when a prompt appears
It used to be that when the `ask` prompt pops up and there has been
enough output already, the ask modal will cover the output the program
has generated so far.

Do an additional "scroll to the bottom" when we open up the modal,
so that the last lines (usually containing the question you need to
answer) are visible.