[FIX] Scroll output to bottom when a prompt appears (#4236)
It used to be that when the `ask` prompt pops up and there has been enough output already, the ask modal will cover the output the program has generated so far.

Do an additional "scroll to the bottom" when we open up the modal, so that the last lines (usually containing the question you need to answer) are visible.

It looks like there was an attempt to do this already, but it was added to the wrong event, so I removed it.

## Before

<img width="766" alt="image" src="https://user-images.githubusercontent.com/524162/232127508-370fc2f4-0bb9-45c1-a327-2b6668844bbf.png">

## After

<img width="766" alt="image" src="https://user-images.githubusercontent.com/524162/232127428-aa2a79ef-00f1-4bd6-b71a-64d009086f91.png">

**How to test**

Emit a large amount of text, followed by a prompt. Observe that the last line is visible.