database file for storing common errors, split up statistics Module
I've split up the live statistics module and already existing Statistics Module to keep them separate since they are quite different. This required the class to be registered in app.py as well.
Moreover, there's a json file with some dummy data that generates the common errors in the live dashboard.
This data structure should probably be merged later with the existing database but for now this will serve our purpose