moved collapse button, changed colour
Move the collapse button of the first section to the top. This was because when testing it with people they found that to be more intuitive rather than below. Moreover, I changed the colour to blue rather than green to have a bit of separation to the 'Back to class' button