[FIX] Make the newlines in the quiz text render properly (#4270)
MarkDown by default doesn't consider newlines in the input text significant: multiple lines of text following another add up to a single paragraph.

It's possible to force newlines by ending a line with two spaces.

While no one is going to do this in the YAML, we can use a filter to replace all newlines with two spaces followed by a newline, to put the input into a format for the `|commonmark` filter that will render the newlines correctly.

Also in this PR: get rid of the now unused `quiz.ts`, which should have been removed as part of #4224, but was forgotten.

**How to test**

Find a quiz question where the answers have newlines in them. See that the newlines are there.