better naming for student details html file, generic url args
Renamed the previously named student-space.html file to class-live-student.html to be in line with other naming. Added a function that dynamically builds the url arguments mainly for dashboard display settings.
This avoids lots of annoying code duplication in the html file and makes it much easier to change args