[UI] Adds function to render maximum level (#4284)
**Description**

Adds a new function to render the maximum level reached by the user in a quiz when they access the /hedy path.

**Fixes #4219*

**How to test**
Log in as a one of the students in the database (student1 to student5), and complete a quiz. Then go to the Hedy root and see if the next level is shown!