[UI] Improvements to class customization (#4252)
**Description**

This PR started as a way to address the issues raised in #3976 and #3955.

The improvements that has been made so far:

1. Removes the buttons in favor of a switch to disable the levels.
2. Adds a status badge to show the status of the current level we're editing.
3. _Adds button to add an adventure right in the customizations page._ (It was decided that this feature will be tackled in a future PR)
4. Removes lengthy explanation at the top of the page. 

Now, once we merged #4224 and added HTMX to Hedy, this is an opportunity to remove a large part of the Javascript code that makes this page work.

**Fixes (partially) #3976, #3955**


**How to test**

Todo: add Cypress tests

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Contains one of the PR categories in the name
- [ ] Describes changes in the format above
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!