[FIX] Use 'input' event for level selector dropdown
The level selector dropdown behaves differently in different browsers:

- In Chrome, it opens up and immediately closes.
- In Safari, it doesn't seem to do anything.

This is due to the event we were listening for: `hx-trigger="click"`.

On Chrome, the click event that opens the dropdown immediately triggers
the HTMX request. On Safari, `click` doesn't seem to be triggered at
all.

Change the event to `hx-trigger="input"`; this gets fired as soon as
a value is selected.