[CHORE] Detect .po file problems during the PR build
For reasons we're not quite sure of, sometimes translations get
lost while generating `.po` files. A situation that happens sometimes
is that we end up with the following in a file:

```
msgid "mandatory_mode"
msgstr ""

(...a lot further down...)

\#~ msgid "mandatory_mode"
\#~ msgstr "Mandatory Developer's Mode
```

For some reason the correct translation is commented out and replaced
with an empty string one.

This PR adds a script that will detect this situation and fail the
build on the PR that tries to accidentally introduce this situation.