[FEAT] Display a warning when the user is not logged in
Warn users if they have been spending a notable amount of effort on a
program and they are not logged in, that their work may be lost if they
don't log in.

The rules are currently a multitude, because we don't know what will
work best. If they either:

- Have click "Run" X times; or
- Their program is longer than Y lines; or
- They've spent more than Z minutes on one tab

Then we show the warning.