[FEAT] Display a warning when the user is not logged in (#4296)
Warn users if they have been spending a notable amount of effort on a program and they are not logged in, that their work may be lost if they don't log in.

The rules are currently a multitude, because we don't know what will work best. If they either:

- Have click "Run" X times; or
- Their program is longer than Y lines; or
- They've spent more than Z minutes on one tab

Then we show the warning.

<img width="456" alt="image" src="https://github.com/hedyorg/hedy/assets/524162/1d9a2bc0-88ae-4ddc-a512-6fb8449ecd46">

**How to test**

Be not logged in, hit the Run button a bunch of times.