removed dummy errors, error history is now used to detect misconceptions
some of the code was commented as the legacy system does not work anymore right now. This has to do with the fact that the string of the exception is stored in the db. Translating the string to class is doable however requires the program code etc. which is not stored for every run rather only for one session