[UI] Fix unreadable quiz output (#4335)
The quiz has an `output` field that gets rendered to show the output of a hypothetical program. The output was being rendered as black text on a dark blue background, and barely readable.

Make it readable and treat the text as `<pre>`formatted.

Also change the width of code and output containers to look slightly nicer.

**How to test**

http://localhost:8080/quiz/preview-question/9/5 will show the difference.

Before:

<img width="1721" alt="image" src="https://github.com/hedyorg/hedy/assets/524162/c68de2f9-72be-42e9-bf7d-25a9c52d2557">

After:

<img width="1236" alt="image" src="https://github.com/hedyorg/hedy/assets/524162/14a95ca9-e531-455c-8301-313013143c95">