[FIX] Remove deprecated extensions from `babel.cfg` (#4338)
**Description**

Removes the line: 
```
extensions=jinja2.ext.autoescape,jinja2.ext.with_
```
from `babel.cfg` which caused an error when executing the extracting phase. This happens because those extensions were deprecated and we recently updated flask to a newer version not realizing this.

**Fixes #4337 **

**How to test**

I tested that everything worked again by adding a test key and seeing if it was extracted and compiled.