[FEATURE] Skipping Faulty Code (#4320)
**Feature**
This PR is made for making skipping faulty code possible in Hedy.

**Example**
A simple program written in level 1

![afbeelding](https://github.com/hedyorg/hedy/assets/48225550/fb5bc2eb-b67f-4bed-a1b4-1fbd820ecc26)


Notice that lines 3 and 6 are skipped, the skipped statements are underlined with a red dashed line, these contain a faulty ```print``` and ```echo``` statement, the rest of the program is still executed! When a user clicks on the skipped section the 'original' error message is shown.

When clicking on ```prind Mistake``` :
![afbeelding](https://github.com/hedyorg/hedy/assets/48225550/2ec986ee-001c-4406-a3cd-8208204f5154)

When clicking on ```eko Hello``` :
![afbeelding](https://github.com/hedyorg/hedy/assets/48225550/0e825b2e-f2a2-4181-a0b7-a80fcc696003)

**Bug - FIXED** 
There seems to be a weird bug. After an error message is shown that is not skipped the program keeps returning an Exception instead of skipping faulty code. But as soon as the local server is restarted it works again.

You can replicate this by first execting ```print hello``` in level 5 and then ```prind 'hello'```. it will skip the faulty code but return no source map due to a TypeError exception, after restarting the local server it will work again as intended.

TO-DO:
- [x] Fix bug mentioned above
- [x] Change the way we test handling exceptions
- [x] Change the way we test the source mapper