[FIX] Fixes Hedy's choice programs not being shown (#4400)
**Description**

Changes the assignment of the `filtered_level` variable that goes to the template. This variable was being converted to `string`, but when there was no `level` argument in the query string, and therefore hold  the value of `None` when it was passed to the template it actually was `'None'` a string. This PR fixes that behaviour.

**Fixes #4399 **