[FIX] Fix share button in Programs page (#4402)
**Description**

Restores the `share_program` function that was deleted in #4057, and also adapted it to the way Rico was calling this type of function. Besides that, so this doesn't happen again I also added a test for this feature.

**Fixes #4398**

**How to test**

You can save a program, go to the Programs page and then share and unshare it! You can also run the cypress tests!