[UI] Changes to grid page UI (#4395)
**Description**

Changes the UI of the grid page to make it more appealing, as stated in #4394. This new page looks like this:

![image](https://github.com/hedyorg/hedy/assets/45865185/bae77558-b1f8-4039-9e0d-04225f281c2b)

Also, changes the name of the template that is supposed to be called with HTMX and Jinja Partials, as per our standards.

**Fixes #4394**

**How to test**

Go to the grid page and see the changes.