[FIX?] Remove Skip Line feature (#4413)
Since the skip lines are failing in ways we don't understand yet, and it hinders our ability to ship new features, this PR temporarily comments some features of the skip lines code. But what do you think of removing them in their entirety? It would look better, but in the future, it'd be harder to reincorporate the feature. What do you think?