[UI] Add pagination to profile page (#4397)
**Description**

Adds pagination to the profile page, limiting to 10 programs per page. It also improves the UI by using the same template for programs as the one used in the /explore page, for this I created a new file that contains that template in this way is easier to keep track of future changes to that UI.

**Fixes #4378**

This is how the page looks now. Timon did such a good work with that design of the programs, that I think is a waste not to use it in this page too!:

![image](https://github.com/hedyorg/hedy/assets/45865185/dd1f22e5-b3d3-4595-bf4a-d9fd11edf8d6)