[FIX] Try to fix and log quiz issues better
We have gotten the following reports from the quiz module:

- Some student has gotten Python exceptions while doing the
  quiz, caused by a `list index out of range` error.
- Another student has gotten `400` errors saying `No such question`.

Trying to address the first issue by assuming the error was caused
by the achievement handling code at the end of a quiz: that code was a
little waffly about the return type of `add_single_achievement`; write
code that can deal with either return value.

For the other issue (and also in case we get the first issue wrong),
start logging more details about the quiz state and progress. Right
now we don't have enough information to diagnose this further.