[FIX] Try to fix and log quiz issues better (#4457)
We have gotten the following reports from the quiz module:

- Some student has gotten Python exceptions while doing the quiz, caused by a `list index out of range` error.
- Another student has gotten `400` errors saying `No such question`.

Trying to address the first issue by assuming the error was caused by the achievement handling code at the end of a quiz: that code was a little waffly about the return type of `add_single_achievement`; write code that can deal with either return value.

For the other issue (and also in case we get the first issue wrong), start logging more details about the quiz state and 

**How to test**

Start the server, click around in the Quiz. Observe that fields like `level` and `question` are logged to the console in the giant dictionary:

![2023-09-06 at 14 01](https://github.com/hedyorg/hedy/assets/524162/97576d6c-a204-4036-92b9-1dda50213c7e)