[TESTING] Adds test for Hedy's choice programs (#4460)
**Description**

The Hedy's choice programs were not functioning for a while, so this tests makes sure that if they fail, it won't go unnoticed again.

**Fixes #4407**

**Checklist**
Done? Check if you have it all in place using this list: (mark with x if done)

- [ ] Contains one of the PR categories in the name
- [ ] Describes changes in the format above
- [ ] Links to an existing issue or discussion
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!