[UI] Adds public profile link to account dropdown (#4492)
**Description**

This PR is based on the work of @Ilia-Gershen on #4456. I couldn't commit changes to that branch, so I had to make a new PR.

Apart from the changes made by Ilia, I also added `has_public_profile` as a variable available in all templates, therefore we can access this variable when deciding whether to show or not the My public profile, button.

**Fixes #4445**

**How to test**

If you're a user with a public profile, the option should be shown in the menu!