[FIX] Make explore page faster
The `explore` page was (effectively) downloading the entire content of
multiple indexes (at least `lang` and `public`, if no other criteria
were given), to do a linear filtering operation.

The `lang` index is just not discriminatory enough, but the good news
is that the `public` index is a lot better. Just use the `public` index,
and filter all programs by the criteria that are provided.

Gets the load time of the `explore` page down to ~1s.