[REFACTORING] Moves markers functionality to editor interface (#4508)
**Description**

One of the points in #4341 was to move the markers functionality to the editor interface. This PR does that, also modifying the debugger file in accordance to the new interface. This changes will allow me to keep working on  #4479, as this is a previous step to that change. 

I basically just copy and pasted the `Markers` class into the `HedyAceEditor` class, modifying it when needed. I also added a couple more editor events, since that was needed in order for the debugger to work.


**How to test**

Everything should work as previously.